+++ src/main/java/gameobj/staticobj/Collectable.java
@@ -9,7 +9,7 @@ public abstract class Collectable extends StaticObj {
     /**
      * Constructor for a collectable; utilises StaticObj's constructor.
      */
-    public Collectable(double posX, double posY, int width, int height, PImage sprite) {
+    public Collectable(float posX, float posY, float width, float height, PImage sprite) {
         super(posX, posY, width, height, sprite);
 
         this.collected = false;
